+++ lib/src/ledger/ledger_ble_manager.dart
@@ -1,16 +1,19 @@
-import 'dart:async';
+import "dart:async";
 
-import 'package:ledger_flutter_plus/src/api/connection_manager.dart';
-import 'package:ledger_flutter_plus/src/api/gatt_gateway.dart';
-import 'package:ledger_flutter_plus/src/exceptions/ledger_exception.dart';
-import 'package:ledger_flutter_plus/src/ledger/connection_type.dart';
-import 'package:ledger_flutter_plus/src/ledger/ledger_gatt_gateway.dart';
-import 'package:ledger_flutter_plus/src/ledger/ledger_transformer.dart';
-import 'package:ledger_flutter_plus/src/ledger_interface.dart';
-import 'package:ledger_flutter_plus/src/models/discovered_ledger.dart';
-import 'package:ledger_flutter_plus/src/models/ledger_device.dart';
-import 'package:ledger_flutter_plus/src/operations/ledger_operations.dart';
-import 'package:universal_ble/universal_ble.dart';
+import "package:flutter/foundation.dart";
+import "package:universal_ble/universal_ble.dart";
+
+import "../api/connection_manager.dart";
+import "../api/gatt_gateway.dart";
+import "../exceptions/ledger_exception.dart";
+import "../ledger_interface.dart";
+import "../models/connection_change_event.dart";
+import "../models/discovered_ledger.dart";
+import "../models/ledger_device.dart";
+import "../operations/ledger_operations.dart";
+import "connection_type.dart";
+import "ledger_gatt_gateway.dart";
+import "ledger_transformer.dart";
 
 const _bleMasterTimeout = Duration(seconds: 60);
 const _bleConnectionTimeout = Duration(seconds: 30);
@@ -21,35 +24,39 @@ class LedgerBleConnectionManager extends ConnectionManager {
 
   final PermissionRequestCallback onPermissionRequest;
 
-  final _connectedDevices =
-      <String, ({GattGateway gateway, LedgerDevice device})>{};
-  final _connectionStateControllers =
-      <String, StreamController<BleConnectionState>>{};
+  final _connectedDevices = <String, ({GattGateway gateway, LedgerDevice device})>{};
+  final _connectionStateControllers = <String, StreamController<BleConnectionState>>{};
 
   final List<OnConnectionChange> _connectionChangeListeners = [];
 
-  final _onConnectionChangeController = StreamController<BleConnectionState>();
+  final _onConnectionChangeController = StreamController<BleConnectionChangeEvent>();
   final _statusStateChangesController = StreamController<AvailabilityState>();
 
   LedgerBleConnectionManager({
     required this.onPermissionRequest,
   }) {
     _connectionChangeListeners.add(_handleConnectionChange);
-    UniversalBle.onConnectionChange = (deviceId, isConnected) {
-      // TODO this is not correct cause it doesn't account for deviceId
-      final state = isConnected
-          ? BleConnectionState.connected
-          : BleConnectionState.disconnected;
-      _onConnectionChangeController.add(state);
+    UniversalBle.onConnectionChange = (deviceId, isConnected, err) {
+      final state = isConnected ? BleConnectionState.connected : BleConnectionState.disconnected;
 
-      // Copy the list because we get a ConcurrentModificationError otherwise
-      final connectionChangeListeners = List.from(
-        _connectionChangeListeners,
-        growable: false,
+      _onConnectionChangeController.add(
+        BleConnectionChangeEvent(
+          deviceId: deviceId,
+          newState: state,
+        ),
       );
 
-      for (final listener in connectionChangeListeners) {
-        listener(deviceId, isConnected);
+      // Copy the list because we get a ConcurrentModificationError otherwise
+      final List<OnConnectionChange> connectionChangeListeners = [
+        ..._connectionChangeListeners,
+      ];
+
+      for (final OnConnectionChange listener in connectionChangeListeners) {
+        try {
+          listener(deviceId, isConnected, err);
+        } catch (e) {
+          debugPrint("Error in connection change listener: $e");
+        }
       }
     };
   }
@@ -58,8 +65,7 @@ class LedgerBleConnectionManager extends ConnectionManager {
   Future<void> connect(LedgerDevice device) async {
     if (_disposed) throw LedgerManagerDisposedException(ConnectionType.ble);
 
-    final availabilityState =
-        await UniversalBle.getBluetoothAvailabilityState();
+    final availabilityState = await UniversalBle.getBluetoothAvailabilityState();
     final granted = await onPermissionRequest(availabilityState);
     if (!granted) {
       return;
@@ -68,53 +74,21 @@ class LedgerBleConnectionManager extends ConnectionManager {
     UniversalBle.timeout = _bleMasterTimeout;
 
     try {
-      final Completer<void> deviceConnected = Completer();
-
-      // ignore: prefer_function_declarations_over_variables
-      final OnConnectionChange connChangeListener = (deviceId, isConnected) {
-        if (deviceId == device.id && !deviceConnected.isCompleted) {
-          if (isConnected) {
-            deviceConnected.complete();
-          }
-          //  else if (error != null) {
-          //   deviceConnected.completeError(error);
-          // }
-        }
-      };
-      final deviceConnectedFuture = deviceConnected.future.timeout(
-        _bleConnectionTimeout,
-        onTimeout: () => throw TimeoutException(
-          "deviceConnectedFuture",
-          _bleConnectionTimeout,
-        ),
-      );
-      deviceConnectedFuture
-          .then((_) => _connectionChangeListeners.remove(connChangeListener))
-          .catchError(
-            (_) => _connectionChangeListeners.remove(connChangeListener),
-          );
-
-      _connectionChangeListeners.add(connChangeListener);
-
       try {
-        final connectionState = await UniversalBle //
-                .getConnectionState(device.id)
-            .timeout(
+        final connectionState = await UniversalBle.getConnectionState(
+          device.id,
+        ).timeout(
           _bleCheckConnectionTimeout,
           onTimeout: () => BleConnectionState.disconnected,
         );
         switch (connectionState) {
           case BleConnectionState.connected:
           case BleConnectionState.connecting:
-            deviceConnected.complete();
             break;
           case BleConnectionState.disconnected:
           case BleConnectionState.disconnecting:
-            // DO NOT AWAIT "connect". It seems to be buggy and not actually complete, despite connChangeListener
-            // getting invoked and confirming that the device connected successfully.
-            UniversalBle.connect(device.id).ignore();
+            await UniversalBle.connect(device.id);
         }
-        await deviceConnectedFuture;
       } catch (e) {
         disconnect(device.id).ignore();
         throw EstablishConnectionException(
@@ -155,11 +129,8 @@ class LedgerBleConnectionManager extends ConnectionManager {
     }
   }
 
-  Future<void> _handleConnectionChange(
-      String deviceId, bool isConnected) async {
-    final state = isConnected
-        ? BleConnectionState.connected
-        : BleConnectionState.disconnected;
+  Future<void> _handleConnectionChange(String deviceId, bool isConnected, String? err) async {
+    final state = isConnected ? BleConnectionState.connected : BleConnectionState.disconnected;
     _getOrCreateConnectionStateController(deviceId).add(state);
   }
 
@@ -168,7 +139,7 @@ class LedgerBleConnectionManager extends ConnectionManager {
   ) =>
       _connectionStateControllers.putIfAbsent(
         deviceId,
-        () => StreamController<BleConnectionState>.broadcast(),
+        StreamController<BleConnectionState>.broadcast,
       );
 
   @override
@@ -182,7 +153,7 @@ class LedgerBleConnectionManager extends ConnectionManager {
     final d = _connectedDevices[device.id];
     if (d == null) {
       throw DeviceNotConnectedException(
-        requestedOperation: 'ble_manager: sendOperation',
+        requestedOperation: "ble_manager: sendOperation",
         connectionType: ConnectionType.ble,
       );
     }
@@ -204,9 +175,7 @@ class LedgerBleConnectionManager extends ConnectionManager {
   Stream<AvailabilityState> get statusStateChanges {
     if (_disposed) throw LedgerManagerDisposedException(ConnectionType.ble);
 
-    UniversalBle.onAvailabilityChange = (state) {
-      _statusStateChangesController.add(state);
-    };
+    UniversalBle.onAvailabilityChange = _statusStateChangesController.add;
     return _statusStateChangesController.stream;
   }
 
@@ -218,10 +187,10 @@ class LedgerBleConnectionManager extends ConnectionManager {
   }
 
   @override
-  Stream<BleConnectionState> get deviceStateChanges {
+  Stream<BleConnectionState> deviceStateChanges(String deviceId) {
     if (_disposed) throw LedgerManagerDisposedException(ConnectionType.ble);
 
-    return _onConnectionChangeController.stream;
+    return _onConnectionChangeController.stream.where((e) => e.deviceId == deviceId).map((e) => e.newState);
   }
 
   @override
@@ -241,7 +210,7 @@ class LedgerBleConnectionManager extends ConnectionManager {
       await UniversalBle.disconnect(deviceId);
     }
 
-    _connectionStateControllers.remove(deviceId)?.close();
+    unawaited(_connectionStateControllers.remove(deviceId)?.close());
   }
 
   @override
@@ -249,8 +218,8 @@ class LedgerBleConnectionManager extends ConnectionManager {
     if (_disposed) return;
     _disposed = true;
 
-    _onConnectionChangeController.close();
-    _statusStateChangesController.close();
+    unawaited(_onConnectionChangeController.close());
+    unawaited(_statusStateChangesController.close());
 
     final deviceIds = List<String>.from(_connectedDevices.keys);
     for (final deviceId in deviceIds) {