+++ lib/src/ledger/ledger_operation.dart
@@ -1,4 +1,4 @@
-import 'package:ledger_flutter_plus/src/operations/ledger_operations.dart';
+import "../operations/ledger_operations.dart";
 
-@Deprecated('Use [LedgerSimpleOperation] or [LedgerComplexOperation] instead')
+@Deprecated("Use [LedgerSimpleOperation] or [LedgerComplexOperation] instead")
 abstract class LedgerOperation<T> extends LedgerRawOperation<T> {}